+++ nacc_attribute_deriver/attributes/missingness/modules/uds/missingness_uds.py
@@ -18,22 +18,6 @@ from nacc_attribute_deriver.symbol_table import SymbolTable
 class UDSMissingness(UDSAttributeCollection):
     """Class to handle UDS missingness values."""
 
-    def handle_v4_missingness(self, field: str) -> Optional[int]:
-        """Handles V4 missingness, which in many cases follows the logic:
-
-        If FORMVER=4 and VAR is blank, VAR should = 0
-        else if FORMVER < 4, VAR should be -4
-        """
-        # if value exists, return None so we don't override
-        value = self.uds.get_value(field, str)
-        if value is not None:
-            return None
-
-        if self.formver == 4:
-            return 0
-
-        return INFORMED_MISSINGNESS
-
     def _missingness_uds(self, field: str) -> Optional[int]:
         """Defines general missingness for UDS; -4 if missing."""
         if self.uds.get_value(field, str) is None: