+++ nacc_attribute_deriver/attributes/nacc/modules/uds/form_a2.py
@@ -11,6 +11,7 @@ from nacc_attribute_deriver.attributes.namespace.namespace import (
     SubjectDerivedNamespace,
 )
 from nacc_attribute_deriver.schema.constants import INFORMED_MISSINGNESS
+from nacc_attribute_deriver.schema.errors import AttributeDeriverError
 from nacc_attribute_deriver.symbol_table import SymbolTable
 
 from .helpers.generate_race import generate_race
@@ -34,7 +35,7 @@ class UDSFormA2Attribute(UDSAttributeCollection):
         NOTE: After discussing with RT, we want to return whatever is on the current
         form and don't carry forward. If not specified, return -4.
         """
-        if not self.submitted:
+        if not self.submitted or self.formver == 4:
             return INFORMED_MISSINGNESS
 
         newinf = self.uds.get_value("newinf", int)
@@ -51,3 +52,31 @@ class UDSFormA2Attribute(UDSAttributeCollection):
         )
 
         return result
+
+    def _create_naccincntfq(self) -> int:
+        """Creates NACCINCNTFQ - frequency of contact."""
+        if not self.submitted:
+            return INFORMED_MISSINGNESS
+
+        if self.formver == 4:
+            inlivwth = self.uds.get_value('inlivwth', int)
+            if inlivwth is None:
+                raise AttributeDeriverError(
+                    "INLIVWTH required if A2 submitted")
+
+            if inlivwth == 1:
+                return 8
+
+            # if INLIVWTH == 0 then INCNTFRQ must be defined
+            # so if it is missing then throw an error
+            incntfrq = self.uds.get_value('incntfrq', int)
+            if incntfrq is None:
+                raise AttributeDeriverError(
+                    "If INLIVWTH == 0, INCNTFRQ must be present")
+            return incntfrq
+
+        # V3 and earlier
+        invisits = self.uds.get_value('invisits', int)
+        incalls = self.uds.get_value('incalls', int)
+        if invisits is not None and incalls is not None:
+            return min(invisits, incalls)